pyproject.toml
[project]
name = "telenotes"
version = "0.1.0"
description = "Notes for telegram."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.11.9",
    "pymongo>=4.15.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "telegramify-markdown>=0.5.1",
]
